+++ nominal/experimental/compute/__init__.py
@@ -1,15 +1,7 @@
 from nominal.experimental.compute._buckets import Bucket, batch_compute_buckets, compute_buckets
-from nominal.experimental.compute._derived_datasets import (
-    commit_derived_definition,
-    create_derived_dataset,
-    get_derived_definition,
-)
 
 __all__ = [
     "Bucket",
     "batch_compute_buckets",
-    "commit_derived_definition",
     "compute_buckets",
-    "create_derived_dataset",
-    "get_derived_definition",
 ]
