+++ nominal/experimental/compute_as_code/__init__.py
@@ -1,0 +1,11 @@
+from nominal.experimental.compute_as_code._derived_datasets import (
+    commit_derived_definition,
+    create_derived_dataset,
+    get_derived_definition,
+)
+
+__all__ = [
+    "commit_derived_definition",
+    "create_derived_dataset",
+    "get_derived_definition",
+]
